extract analysis download button
